requirements.txt
streamlit
PyPDF2
langchain
langchain-community
openai
faiss-cpu
tiktoken
python-docx
pytesseract
Pillow